Added functions to easily clone and upload a new version.
The cloneVersion function creates a directory containing symlinks to cached
files; these symlinks are then parsed by prepareDirectoryUpload to be converted
into upload links in startUpload. The idea is to generate a directory for a new
version of an asset, which users can naturally modify to do whatever changes
they need to; but any non-modified files in the directory will remain as
symlinks and be automatically converted into upload links.