Sanitize multiple slashes, Windows backslashes in paths.
This ensures that the paths can be properly used as bucket keys during upload,
downloads of individual files, and conversion of symlinks. Note that the
converse is not true; we assume that '/' is an acceptable file separator on all
platforms, and we don't try to consider arbitrary separators on weird systems.