Avoid doing anything with symbolic links on Windows. (#3)
It seems like symlink support is pretty unreliable on Windows, so cloneVersion
now just falls back to creating hard links or copies if it's on Windows. This
makes prepareDirectoryUpload less efficient as it can't create upload links,
but the backend should end up deduplicating on MD5 sums, so it's okay.

All symlink-related tests are now disabled on Windows as a result.